+++ nextcloudpi.sh
@@ -29,7 +29,7 @@ install()
   apt-get update
   $APTINSTALL dialog whiptail
   mkdir -p $CONFDIR
-  ln -s /usr/local/bin/nextcloudpi-config /usr/local/bin/ncp-config
+  ln -sf /usr/local/bin/nextcloudpi-config /usr/local/bin/ncp-config
 
   # include option in raspi-config (only Raspbian)
   test -f /usr/bin/raspi-config && {
@@ -88,9 +88,22 @@ Listen 4443
 
 </Directory>
 EOF
+
+  cat > /etc/apache2/sites-available/ncp-http-redirect.conf <<'EOF'
+<VirtualHost _default_:4443>
+  DocumentRoot /var/www/nextcloud
+  <IfModule mod_rewrite.c>
+    RewriteEngine On
+    RewriteCond %{HTTPS} !=on
+    RewriteRule ^/?(.*) https://%{SERVER_NAME}/$1 [R,L]
+  </IfModule>
+</VirtualHost>
+EOF
+
   $APTINSTALL libapache2-mod-authnz-external pwauth
   a2enmod authnz_external authn_core auth_basic
   a2ensite ncp
+  a2ensite ncp-http-redirect
 
   ## NCP USER FOR AUTHENTICATION
   useradd $WEBADMIN
@@ -107,20 +120,10 @@ DIR=/usr/local/etc/nextcloudpi-config.d
 test -f $DIR/$1 || { echo "File not found"; exit 1; }
 source /usr/local/etc/library.sh
 cd $DIR
-touch /run/ncp.log
-chmod 640 /run/ncp.log
-chown root:www-data /run/ncp.log
-launch_script $1 &> /run/ncp.log
-RET=$?
-
-# clean log for the next PHP backend call to start clean,
-# but wait until everything from current execution is read
-sleep 0.5 && echo "" > /run/ncp.log
-
-exit $RET
+launch_script $1
 EOF
   chmod 700 /home/www/ncp-launcher.sh
-  echo "www-data ALL = NOPASSWD: /home/www/ncp-launcher.sh , /sbin/halt" >> /etc/sudoers
+  echo "www-data ALL = NOPASSWD: /home/www/ncp-launcher.sh , /sbin/halt, /sbin/reboot" >> /etc/sudoers
 
   # NEXTCLOUDPI AUTO TRUSTED DOMAIN
   mkdir -p /usr/lib/systemd/system